+++ changelog.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - advapi.RegisterServiceCtrlHandler
 - advapi.SetServiceStatus
 - sample lsc.py
+- advapi.GetTokenInformation
 
 ### Changed